[ADD] API 분리
