+++ API/urls.py
@@ -3,4 +3,5 @@ from . import views
 
 urlpatterns = [
     # path('data_insert', views.data_insert, name='data_insert'),
+    path('view/<int:id>', views.user_view_in),
 ]
