+++ main_page/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     path('', views.main, name='main'),
-    path('view/<int:id>', views.user_view_in),
 ]
